backend/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
coverage==5.5
Django==3.1.5
django-cors-headers==3.6.0
django-crontab==0.7.1
django-log-viewer==1.1.1
django-mongolog==0.9.4
django-oauth-toolkit==1.3.3
djangorestframework==3.12.2
et-xmlfile==1.1.0
idna==2.10
numpy==1.20.1
oauthlib==3.1.0
openpyxl==3.0.7
Pillow==8.3.0
psycopg2-binary==2.8.6
PyJWT==2.0.1
pymongo==3.12.0
python-dotenv==0.15.0
pytz==2020.5
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.2
python-dateutil==2.8.2
omise==0.9.0